+++ druid-shell/src/platform/windows/util.rs
@@ -26,9 +26,9 @@ use std::slice;
 use lazy_static::lazy_static;
 use winapi::ctypes::c_void;
 use winapi::shared::guiddef::REFIID;
-use winapi::shared::minwindef::{HMODULE, UINT};
+use winapi::shared::minwindef::{HMODULE, UINT, BOOL};
 use winapi::shared::ntdef::{HRESULT, LPWSTR};
-use winapi::shared::windef::{HMONITOR, RECT};
+use winapi::shared::windef::{HMONITOR, RECT, HWND};
 use winapi::shared::winerror::SUCCEEDED;
 use winapi::um::fileapi::{CreateFileA, GetFileType, OPEN_EXISTING};
 use winapi::um::handleapi::INVALID_HANDLE_VALUE;
@@ -128,10 +128,12 @@ pub(crate) fn recti_to_rect(rect: RECT) -> Rect {
 }
 
 // Types for functions we want to load, which are only supported on newer windows versions
-// from shcore.dll
-type GetDpiForSystem = unsafe extern "system" fn() -> UINT;
-type GetDpiForMonitor = unsafe extern "system" fn(HMONITOR, MONITOR_DPI_TYPE, *mut UINT, *mut UINT);
 // from user32.dll
+type GetDpiForSystem = unsafe extern "system" fn() -> UINT;
+type GetDpiForWindow = unsafe extern "system" fn(HWND) -> UINT;
+type SetProcessDpiAwarenessContext = unsafe extern "system" fn(winapi::shared::windef::DPI_AWARENESS_CONTEXT) -> BOOL;
+// from shcore.dll
+type GetDpiForMonitor = unsafe extern "system" fn(HMONITOR, MONITOR_DPI_TYPE, *mut UINT, *mut UINT);
 type SetProcessDpiAwareness = unsafe extern "system" fn(PROCESS_DPI_AWARENESS) -> HRESULT;
 type DCompositionCreateDevice2 = unsafe extern "system" fn(
     renderingDevice: *const IUnknown,
@@ -144,6 +146,8 @@ type CreateDXGIFactory2 =
 #[allow(non_snake_case)] // For member fields
 pub struct OptionalFunctions {
     pub GetDpiForSystem: Option<GetDpiForSystem>,
+    pub GetDpiForWindow: Option<GetDpiForWindow>,
+    pub SetProcessDpiAwarenessContext: Option<SetProcessDpiAwarenessContext>,
     pub GetDpiForMonitor: Option<GetDpiForMonitor>,
     pub SetProcessDpiAwareness: Option<SetProcessDpiAwareness>,
     pub DCompositionCreateDevice2: Option<DCompositionCreateDevice2>,
@@ -198,6 +202,8 @@ fn load_optional_functions() -> OptionalFunctions {
 
     let mut GetDpiForSystem = None;
     let mut GetDpiForMonitor = None;
+    let mut GetDpiForWindow = None;
+    let mut SetProcessDpiAwarenessContext = None;
     let mut SetProcessDpiAwareness = None;
     let mut DCompositionCreateDevice2 = None;
     let mut CreateDXGIFactory2 = None;
@@ -213,6 +219,8 @@ fn load_optional_functions() -> OptionalFunctions {
         log::info!("No user32.dll");
     } else {
         load_function!(user32, GetDpiForSystem, "10");
+        load_function!(user32, GetDpiForWindow, "10");
+        load_function!(user32, SetProcessDpiAwarenessContext, "10");
     }
 
     if !dcomp.is_null() {
@@ -225,6 +233,8 @@ fn load_optional_functions() -> OptionalFunctions {
 
     OptionalFunctions {
         GetDpiForSystem,
+        GetDpiForWindow,
+        SetProcessDpiAwarenessContext,
         GetDpiForMonitor,
         SetProcessDpiAwareness,
         DCompositionCreateDevice2,